Web-gym/requirements.txt
django==4.0.1
Pillow==9.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
geopy==2.3.0
geocoder==1.38.1
python-dateutil==2.8.2
django-cors-headers==3.13.0